Refactor: relocate Header component to feature directory and update imports
