+++ src/app/home/page.tsx
@@ -1,4 +1,4 @@
-import Header from '@/components/Header';
+import Header from '@/feature/Header/Header';
 import { StartTrainingButton } from '@/feature/StartTrainingButton/StartTrainingButton';
 import React from 'react';
